Used float instead for checking percentages

+++ UnityProject/Assets/CommonCore/Data/RangeDataWithSingleString.cs
@@ -3,8 +3,8 @@ using System;
 namespace Common {
     [Serializable]
     public struct RangeDataWithSingleString {
-        public int min;
-        public int max;
+        public float min;
+        public float max;
         public string stringId;
-    } 
+    }
 }
